busybox: add a DEPENDS on libpam when pam is selected in DISTRO_FEATURES
When 'pam' is in DISTRO_FEATURES and CONFIG_PAM is selected in busybox
the target-libpam is required to be installed before busybox can be
built. Add a build-time dependency for busybox on libpam to guarantee
this.
If CONFIG_PAM is not selected in busybox's .config, this dependency
would rather be unnecessary, but it doesn't hurt to have it anyhow.