package.bbclass: skip non-existant files/directories when doing fixup_perms
The global fs-perms.txt may contain references to user/group ids that
are created during the build process. Currently fixup_perms chokes on
entries in fs-perms.txt that have unknown (on the target) user/group
names because it tries to resolve the uid/gid immediately when
scanning the fs-perms.txt file.

This patch makes it check for the existence of the corresponding file
before trying to resolve the uid/gid.

If a file does not exist in the rootfs (yet), the corresponding
uid/gid specified in fs-perms.txt is meaningless and will be skipped.

If a file has been installed that requires certain uid/gid settings,
the corresponding user/group names should have been added to
etc/passwd|etc/group in the same (or an earlier) build step and the
uid/gid lookup will succeed.

This makes it possible to specify arbitrary files with arbitrary
user/group names in fs-perms.txt without breaking the build no matter
at what build step those files are actually installed.