grub2: Add CVE whitelist entries for issues fixed in 2.06
We're using a pre-release version of 2.06 so these issues are fixed but
continue to show up in the checks since it is pre-2.06 and the CPE
entries are "before but excluding 2.06".

Adding these will clean up CVE reports until the 2.06 release comes out.

(From OE-Core rev: 07451418e8ffef608e05b981bf7516bef5450d49)

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>